% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grond-package.R
\docType{package}
\name{grond-package}
\alias{grond}
\alias{grond-package}
\title{grond: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Dave Gruenewald \email{dave.gruenewald@posit.co}

}
\keyword{internal}
