% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lotr_data}
\alias{lotr_data}
\title{Quotes from the Lord of the Rings}
\format{
A data frame with 65 rows and 2 variables:
\describe{
\item{quote}{character string from book}
\item{character}{Common name of persons from Lord of the Rings}
}
}
\usage{
lotr_data
}
\description{
A dataset containing the quotes of main characters throughout the Lord of the
Rings trilogy
}
\keyword{datasets}
