% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote_lotr.R
\name{quote_lotr}
\alias{quote_lotr}
\title{Quoting Lord of the Rings}
\usage{
quote_lotr(character)
}
\arguments{
\item{character}{Common name for main characters in Lord of the Rings}
}
\value{
character string from Lord of the Rings
}
\description{
Quoting Lord of the Rings
}
\examples{
quote_lotr("Gandalf")
}
